[wpilib] Deprecate MotorController
The drive classes no longer need this interface, and it rigidly couples
us to motor controller vendors who only update and do a major release of
their vendordep once per year.

We were originally going to do this in #6053, but we wanted to gather
feedback on the alternate approach during a season first (as far as I
know, we received none).